refactor(ui): route-based settings items
