+++ ui/apps/settings-app/src/components/option-apps.tsx
@@ -3,7 +3,7 @@ import DS from '@akashaproject/design-system';
 import { StyledIcon, StyledText } from '../styles';
 import { IAppsOption } from '../interfaces';
 
-const { Box, Checkbox } = DS;
+const { Box, BasicCardBox, Checkbox } = DS;
 
 const AppsOption: React.FC<IAppsOption> = props => {
   const {
@@ -20,59 +20,65 @@ const AppsOption: React.FC<IAppsOption> = props => {
   } = props;
 
   return (
-    <>
-      <Box
-        direction="row"
-        pad="medium"
-        align="center"
-        border={{ side: 'bottom', color: 'lightBorder' }}
-      >
-        <StyledIcon type="chevronLeft" onClick={OnChevronLeftClick} />
-        <StyledText weight="bold" size="large" margin={{ vertical: '0', horizontal: 'auto' }}>
-          {titleLabel}
-        </StyledText>
-      </Box>
-      <Box
-        direction="row"
-        pad="medium"
-        justify="between"
-        align="center"
-        border={{ side: 'bottom', color: 'lightBorder' }}
-      >
-        <StyledText weight="bold">{autoUpdatesLabel}</StyledText>
-        <Box pad={{ top: 'small' }}>
-          <Checkbox checked={checkedAutoUpdates} onChange={onAutoUpdatesChange} toggle={true} />
+    <Box direction="column" gap="small">
+      <BasicCardBox>
+        <Box
+          direction="row"
+          pad="medium"
+          align="center"
+          border={{ side: 'bottom', color: 'lightBorder' }}
+        >
+          <StyledIcon type="chevronLeft" onClick={OnChevronLeftClick} />
+          <StyledText weight="bold" size="large" margin={{ vertical: '0', horizontal: 'auto' }}>
+            {titleLabel}
+          </StyledText>
         </Box>
-      </Box>
-      <Box
-        pad={{ top: 'medium', horizontal: 'medium', bottom: 'large' }}
-        justify="center"
-        align="start"
-        border={{ side: 'bottom', color: 'lightBorder' }}
-      >
-        <StyledText>{autoUpdatesInfo}</StyledText>
-      </Box>
-      <Box
-        direction="row"
-        pad="medium"
-        justify="between"
-        align="center"
-        border={{ side: 'bottom', color: 'lightBorder' }}
-      >
-        <StyledText weight="bold">{dataAnalyticsLabel}</StyledText>
-        <Box pad={{ top: 'small' }}>
-          <Checkbox checked={checkedDataAnalytics} onChange={onDataAnalyticsChange} toggle={true} />
+        <Box
+          direction="row"
+          pad="medium"
+          justify="between"
+          align="center"
+          border={{ side: 'bottom', color: 'lightBorder' }}
+        >
+          <StyledText weight="bold">{autoUpdatesLabel}</StyledText>
+          <Box pad={{ top: 'small' }}>
+            <Checkbox checked={checkedAutoUpdates} onChange={onAutoUpdatesChange} toggle={true} />
+          </Box>
         </Box>
-      </Box>
-      <Box
-        pad={{ top: 'medium', horizontal: 'medium', bottom: 'large' }}
-        justify="center"
-        align="start"
-        border={{ side: 'bottom', color: 'lightBorder' }}
-      >
-        <StyledText>{dataAnalyticsinfo}</StyledText>
-      </Box>
-    </>
+        <Box
+          pad={{ top: 'medium', horizontal: 'medium', bottom: 'large' }}
+          justify="center"
+          align="start"
+          border={{ side: 'bottom', color: 'lightBorder' }}
+        >
+          <StyledText>{autoUpdatesInfo}</StyledText>
+        </Box>
+        <Box
+          direction="row"
+          pad="medium"
+          justify="between"
+          align="center"
+          border={{ side: 'bottom', color: 'lightBorder' }}
+        >
+          <StyledText weight="bold">{dataAnalyticsLabel}</StyledText>
+          <Box pad={{ top: 'small' }}>
+            <Checkbox
+              checked={checkedDataAnalytics}
+              onChange={onDataAnalyticsChange}
+              toggle={true}
+            />
+          </Box>
+        </Box>
+        <Box
+          pad={{ top: 'medium', horizontal: 'medium', bottom: 'large' }}
+          justify="center"
+          align="start"
+          border={{ side: 'bottom', color: 'lightBorder' }}
+        >
+          <StyledText>{dataAnalyticsinfo}</StyledText>
+        </Box>
+      </BasicCardBox>
+    </Box>
   );
 };
